+++ plugins/plugin-feed/src/node/feedPlugin.ts
@@ -4,9 +4,9 @@ import { isLinkHttp, removeEndingSlash } from 'vuepress/shared'
 import { colors } from 'vuepress/utils'
 import type { FeedPluginOptions } from '../typings/index.js'
 import { addFeedLinks } from './addFeedLinks.js'
-import { getFeedFiles } from './getFeed.js'
-import { getAtomTemplates, getRSSTemplates } from './getTemplate.js'
-import { getFeedOptions } from './options.js'
+import { getFeedFiles } from './getFeedFiles.js'
+import { getFeedOptions } from './getFeedOptions.js'
+import { getAtomTemplates, getRSSTemplates } from './getTemplates.js'
 import { writeFiles } from './output.js'
 import { FEED_GENERATOR, logger } from './utils/index.js'
