+++ plugins/plugin-feed/src/node/getTemplates.ts
@@ -2,8 +2,8 @@ import { entries } from '@vuepress/helper/node'
 import { ensureEndingSlash } from 'vuepress/shared'
 import { fs, getDirname, path } from 'vuepress/utils'
 import type { FeedConfig } from '../typings/index.js'
-import type { ResolvedFeedOptionsMap } from './options.js'
-import { getFilename } from './options.js'
+import { getFeedFilenames } from './getFeedFilenames.js'
+import type { ResolvedFeedOptionsMap } from './getFeedOptions.js'
 
 const __dirname = getDirname(import.meta.url)
 
@@ -30,7 +30,7 @@ export const getAtomTemplates = (
     // write template
     .map(([localePath, localeOptions]) => {
       const { atomXslTemplate = DEFAULT_ATOM_XML_TEMPLATE } = localeOptions
-      const { atomXslFilename } = getFilename(localeOptions, localePath)
+      const { atomXslFilename } = getFeedFilenames(localeOptions, localePath)
 
       return [atomXslFilename, atomXslTemplate]
     })
@@ -45,7 +45,7 @@ export const getRSSTemplates = (
     .map(([localePath, localeOptions]) => {
       const { rssXslTemplate = DEFAULT_RSS_XML_TEMPLATE } = localeOptions
 
-      const { rssXslFilename } = getFilename(localeOptions, localePath)
+      const { rssXslFilename } = getFeedFilenames(localeOptions, localePath)
 
       return [rssXslFilename, rssXslTemplate]
     })
