--- plugins/plugin-feed/src/node/options.ts
@@ -1,209 +1,0 @@
-import {
-  dateSorter,
-  fromEntries,
-  isArray,
-  isFunction,
-  keys,
-} from '@vuepress/helper/node'
-import type { GitData } from '@vuepress/plugin-git'
-import type { App, Page } from 'vuepress/core'
-import { isLinkHttp, removeLeadingSlash } from 'vuepress/shared'
-import type {
-  BaseFeedPluginOptions,
-  FeedChannelOption,
-  FeedPluginOptions,
-} from '../typings/index.js'
-import { getUrl } from './utils/index.js'
-
-export interface ResolvedFeedOptions
-  extends Omit<
-      BaseFeedPluginOptions,
-      'sorter' | 'filter' | 'preservedElements'
-    >,
-    Required<Pick<BaseFeedPluginOptions, 'sorter' | 'filter'>> {
-  hostname: string
-  isPreservedElement: (tagName: string) => boolean
-}
-
-export type ResolvedFeedOptionsMap = Record<string, ResolvedFeedOptions>
-
-export const getFeedOptions = (
-  { siteData }: App,
-  options: FeedPluginOptions,
-): ResolvedFeedOptionsMap =>
-  fromEntries(
-    keys({
-      // root locale must exists
-      '/': {},
-      ...siteData.locales,
-    }).map((localePath) => {
-      const preservedElements =
-        options.locales?.[localePath]?.preservedElements ||
-        options.preservedElements
-      const { hostname, devServer, locales, ...rest } = options
-
-      return [
-        localePath,
-        {
-          // default values
-          filter: ({ frontmatter, filePathRelative }: Page): boolean =>
-            !(
-              frontmatter.home ||
-              !filePathRelative ||
-              frontmatter.article === false ||
-              frontmatter.feed === false
-            ),
-          sorter: (
-            pageA: Page<{ git?: GitData }, Record<string, never>>,
-            pageB: Page<{ git?: GitData }, Record<string, never>>,
-          ): number =>
-            dateSorter(
-              pageA.data.git?.createdTime
-                ? new Date(pageA.data.git?.createdTime)
-                : pageA.frontmatter.date,
-              pageB.data.git?.createdTime
-                ? new Date(pageB.data.git?.createdTime)
-                : pageB.frontmatter.date,
-            ),
-
-          ...rest,
-          ...options.locales?.[localePath],
-
-          // make sure these are not overrode
-          hostname,
-          isPreservedElement: isArray(preservedElements)
-            ? (tagName: string): boolean =>
-                preservedElements.some((item) =>
-                  item instanceof RegExp
-                    ? item.test(tagName)
-                    : item === tagName,
-                )
-            : isFunction(preservedElements)
-              ? preservedElements
-              : (): boolean => false,
-        } as ResolvedFeedOptions,
-      ]
-    }),
-  )
-
-export const getFeedChannelOption = (
-  app: App,
-  options: FeedPluginOptions,
-  localePath = '',
-): FeedChannelOption => {
-  const { base } = app.options
-  const { title, description, lang, locales } = app.siteData
-  const {
-    channel: { icon: channelIcon, image: channelImage, ...channel } = {},
-    hostname,
-    icon,
-    image,
-  } = options
-  const authorName = isArray(options.channel?.author)
-    ? options.channel?.author[0]?.name
-    : options.channel?.author?.name
-
-  const defaultChannelOption: FeedChannelOption = {
-    title: locales[localePath]?.title || title || locales['/']?.title || '',
-    link: getUrl(hostname, base, localePath),
-    description:
-      locales[localePath]?.description ||
-      description ||
-      locales['/']?.description ||
-      '',
-    language: locales[localePath]?.lang || lang,
-    copyright: authorName ? `Copyright by ${authorName}` : '',
-    pubDate: new Date(),
-    lastUpdated: new Date(),
-    ...(icon
-      ? { icon: isLinkHttp(icon) ? icon : getUrl(hostname, base, icon) }
-      : {}),
-    ...(image
-      ? { image: isLinkHttp(image) ? image : getUrl(hostname, base, image) }
-      : {}),
-  }
-
-  return {
-    ...defaultChannelOption,
-    ...channel,
-    ...(channelIcon
-      ? {
-          icon: isLinkHttp(channelIcon)
-            ? channelIcon
-            : getUrl(hostname, base, channelIcon),
-        }
-      : {}),
-    ...(channelImage
-      ? {
-          image: isLinkHttp(channelImage)
-            ? channelImage
-            : getUrl(hostname, base, channelImage),
-        }
-      : {}),
-  }
-}
-
-export const getFilename = (
-  options: ResolvedFeedOptions,
-  prefix = '/',
-): Required<
-  Pick<
-    FeedPluginOptions,
-    | 'atomOutputFilename'
-    | 'atomXslFilename'
-    | 'jsonOutputFilename'
-    | 'rssOutputFilename'
-    | 'rssXslFilename'
-  >
-> => ({
-  atomOutputFilename: `${removeLeadingSlash(prefix)}${
-    options.atomOutputFilename || 'atom.xml'
-  }`,
-  atomXslFilename: `${removeLeadingSlash(prefix)}${
-    options.atomXslFilename || 'atom.xsl'
-  }`,
-
-  jsonOutputFilename: `${removeLeadingSlash(prefix)}${
-    options.jsonOutputFilename || 'feed.json'
-  }`,
-  rssOutputFilename: `${removeLeadingSlash(prefix)}${
-    options.rssOutputFilename || 'rss.xml'
-  }`,
-  rssXslFilename: `${removeLeadingSlash(prefix)}${
-    options.rssXslFilename || 'rss.xsl'
-  }`,
-})
-
-export interface FeedLinks {
-  localePath: string
-  atom: string
-  atomXsl: string
-  json: string
-  rss: string
-  rssXsl: string
-}
-
-export const getFeedLinks = (
-  app: App,
-  options: ResolvedFeedOptions,
-  localePath: string,
-): FeedLinks => {
-  const { base } = app.options
-  const { hostname } = options
-  const {
-    atomOutputFilename,
-    atomXslFilename,
-    jsonOutputFilename,
-    rssOutputFilename,
-    rssXslFilename,
-  } = getFilename(options, localePath)
-
-  return {
-    localePath,
-    atom: getUrl(hostname, base, atomOutputFilename),
-    atomXsl: getUrl(hostname, base, atomXslFilename),
-    json: getUrl(hostname, base, jsonOutputFilename),
-    rss: getUrl(hostname, base, rssOutputFilename),
-    rssXsl: getUrl(hostname, base, rssXslFilename),
-  }
-}
